chapter11/gpu/app/requirements.txt
flask==2.3.3
prometheus-client==0.17.1
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
gunicorn==21.2.0